microclimate-crop-advisor/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
openai==1.3.0
python-dotenv==1.0.0